fix: Stop depending on dependency matrix for jx version to release
The dependency matrix can get messed up such that the ML builder
dependency overrules any others, meaning that if the ML builders are
stale for some reason, the jx version is never actually "released". So
let's instead rely on the auto-updating magic of the builder versions
in `jenkins-x.yml` and `jx version --short`.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>